kafka_consumer/requirements-freeze.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
Bottleneck @ file:///croot/bottleneck_1731058641041/work
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.3
fastapi==0.115.12
fastapi-utils==0.8.0
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h5py @ file:///croot/h5py_1750862990877/work
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kafka-python==2.0.5
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
maxminddb==2.7.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_dtypes==0.5.3
mpi4py @ file:///croot/mpi4py_1749458488241/work
multidict==6.6.4
mysql-connector-python==9.2.0
namex==0.1.0
numexpr @ file:///croot/numexpr_1752521720723/work
numpy==2.3.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pika==1.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
psutil==5.9.8
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1752135852232/work
pyudev==0.24.3
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.1.0
scikit-learn==1.6.1
scipy==1.16.1
setuptools==78.1.1
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.23.2
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
